move static js/css to static folder
