+++ src/server/web/endpoint.rs
@@ -1,4 +1,4 @@
-use std::collections::HashMap;
+use std::{borrow::Cow, collections::HashMap};
 
 use either::Either;
 use maud::{html, Markup, PreEscaped, DOCTYPE};
@@ -366,232 +366,12 @@ fn render_feed(mut feed: Feed) -> Markup {
   }
 }
 
-fn inline_styles() -> &'static str {
-  r#"
-  .icon {
-    transition: all 0.2s;
-  }
-
-  .post-entry {
-    margin-left: 0 !important;
-    margin-right: 0 !important;
-
-    .icon-container {
-      display: inline-flex;
-      align-self: center;
-      margin: 0 0.2rem;
-    }
-
-    &[data-folded="false"] {
-      .fold-icon > .icon {
-        transform: rotate(90deg);
-      }
-    }
-    &[data-folded="true"] {
-      header {
-        border: 0 !important;
-        margin-bottom: 0 !important;
-        padding-bottom: 0 !important;
-      }
-
-      section, footer {
-        display: none;
-      }
-    }
-
-    .entry-content {
-      display: none;
-    }
-    &[data-display-mode="rendered"] {
-      .entry-content.rendered {
-        display: block;
-        overflow-x: scroll;
-      }
-    }
-    &[data-display-mode="raw"] {
-      .raw-icon > .icon {
-        color: var(--active);
-      }
-      .entry-content.raw {
-        display: block;
-      }
-    }
-    &[data-display-mode="json"] {
-      .json-icon > .icon {
-        color: var(--active);
-      }
-      .entry-content.json {
-        display: block;
-      }
-    }
-  }
-
-  .source {
-    font-family: monospace;
-  }
-
-  @keyframes rotation {
-    from { transform: rotate(0deg); }
-    to   { transform: rotate(360deg); }
-  }
-
-  .loading {
-    display: none;
-    position: absolute;
-    right: 1rem;
-    align-items: center;
-    height: 100%;
-
-    &.htmx-request {
-      display: flex;
-      animation: rotation 2s infinite linear;
-    }
-  }
-
-  .filter-item {
-    position: relative;
-
-    > var {
-      cursor: pointer;
-    }
-
-    > .filter-definition, .filter-link {
-      display: none;
-    }
-
-    &:hover > .filter-link {
-      display: inline-block;
-      border-top: 1px solid var(--bd-muted);
-      border-bottom: 1px solid var(--bd-muted);
-      margin-left: 0.2rem;
-      width: 15rem;
-      vertical-align: middle;
-    }
-
-    &:hover > .filter-definition {
-      display: block;
-      position: absolute;
-      left: 15rem;
-      top: 0;
-      z-index: 1;
-      border: 1px solid var(--bd-muted);
-      border-radius: var(--bd-radius);
-      box-shadow: 1px 2px 3px var(--bd-muted);
-    }
-  }
-
-  .source-control {
-    background-color: var(--bg-muted);
-    padding: 1rem;
-    border-radius: var(--bd-radius);
-    display :flex;
-    position: relative;
-    align-items: center;
-  }
-  .source-template-container {
-    display: flex;
-    position: relative;
-    align-items: baseline;
-    flex-wrap: wrap;
-
-    .source-template-placeholder {
-      width: auto;
-      display: inline-block;
-    }
-  }
-
-
-  main.feed-section {
-    background-color: var(--bg-muted);
-    padding: 1.5rem;
-    border-radius: var(--bd-radius);
-  }
-
-  .header-bar {
-    margin: 1rem 0 !important;
-    padding-bottom: 1rem;
-    border-bottom: 1.5px dotted;
-    display: flex;
-    align-items: center;
-    height: 2rem;
-
-    > h2 {
-      flex: 1;
-      margin-bottom: 0;
-    }
-
-    > .back-button {
-      margin-right: 2rem;
-
-      a:hover {
-        color: var(--bg-accent);
-      }
-    }
-
-    > .copy-button {
-      > svg {
-        vertical-align: middle;
-      }
-    }
-  }
-
-  .filter-item {
-    var.filter-name[data-enabled="0"] {
-      color: var(--muted);
-      background-color: var(--bg-muted);
-    }
-  }
-
-  pre[class*="language"] {
-    margin: 0 !important;
-  }
-  "#
+fn inline_styles() -> Cow<'static, str> {
+  super::Asset::get_content("endpoint.css")
 }
 
-fn inline_script() -> &'static str {
-  r#"
-  function toggleFold(element) {
-    const article = element.closest("article");
-    article.dataset.folded = article.dataset.folded === "false";
-  }
-
-  function toggleRaw(element) {
-    const article = element.closest("article");
-    article.dataset.folded = "false";
-    article.dataset.displayMode =
-      article.dataset.displayMode === "raw" ? "rendered" : "raw";
-  }
-
-  function toggleJson(element) {
-    const article = element.closest("article");
-    article.dataset.folded = "false";
-    article.dataset.displayMode =
-      article.dataset.displayMode === "json" ? "rendered" : "json";
-  }
-
-  function gatherFilterSkip() {
-    const skipped = [...document.querySelectorAll(".filter-item > var")].filter(x=>!+x.dataset.enabled).map(x => x.dataset.index).join(",");
-    if (skipped === "") {
-      return {};
-    } else {
-      return {filter_skip: skipped};
-    }
-  }
-
-  function fillEntryContent(id) {
-    const parent = document.getElementById(id);
-    const shadowRoot = parent.attachShadow({mode: 'open'});
-    const content = parent.querySelector("template").innerHTML;
-    parent.innerHTML = "";
-    shadowRoot.innerHTML = content;
-  }
-
-  function copyToClipboard() {
-    const url = window.location.href.replace(/\/_\/endpoint\//, "/");
-    navigator.clipboard.writeText(url);
-    alert("Copied: " + url);
-  }
-  "#
+fn inline_script() -> Cow<'static, str> {
+  super::Asset::get_content("endpoint.js")
 }
 
 // requires the container to have a `display: flex` style
