+++ static/list.css
@@ -1,0 +1,8 @@
+.header-bar {
+  margin: 1rem 0 !important;
+  padding-bottom: 1rem;
+  border-bottom: 1.5px dotted;
+  display: flex;
+  align-items: center;
+  height: 2rem;
+}
