+++ static/login.css
@@ -1,0 +1,22 @@
+body {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+form {
+  margin-top: 15px;
+  margin-bottom: 20px;
+  display: flex;
+  flex-direction: column;
+  width: 200px;
+  gap: 0.5rem;
+}
+
+.hidden {
+  display: none;
+}
+
+p#message {
+  color: red;
+}
